change where conversion to int happens

+++ lab/move.py
@@ -67,11 +67,11 @@ def map_environment():
 
     distance = 20
     alpha = -60
-    x = int(environment_size + (distance * math.sin(alpha)))
-    y = int(distance * math.cos(alpha))
+    x = environment_size + (distance * math.sin(alpha))
+    y = distance * math.cos(alpha)
 
     print("X: %s, Y: %s", x, y)
-    environment[x, y] = 1
+    environment[int(x), int(y)] = 1
 
 
 if __name__ == '__main__':
